requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
qrcode==7.4.2
Pillow==10.0.0
cryptography==41.0.0
pyotp==2.9.0
python-dotenv==1.0.0
werkzeug==2.3.0
